+++ src/pages/sign_in.js
@@ -2,8 +2,8 @@ import React, { useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
 import UserContext from '../context/user_context'; 
 import Form from '../components/form';
-import { API_ENDPOINTS } from '../utilities/api_config';
-import { fetchCsrfToken, post } from '../utilities/api_service';
+import { API_ENDPOINTS } from '../utilities/api-config';
+import { fetchCsrfToken, post } from '../utilities/api-service';
 import './general_page.css';
 
 const SignIn = () => {
